Modify message publishing to be more generic
Changed the publisher to not be kv events specific. Instead make it
more flexible to handle other message types.
